Replace f64 with f32 in worldgen code
The extra precision granted by f64 is unnecessary within a single node,
and when working across nodes, and f64 is insufficient, and extra steps
are required (such as the exponent field in Plane). These extra steps
will make f32 sufficient, so f64 will usually not be necessary.

+++ common/src/math.rs
@@ -1150,15 +1150,15 @@ mod tests {
 
     #[test]
     fn distance_commutative() {
-        let p = MPoint::new_unchecked(-1.0, -1.0, 0.0, 3.0f64.sqrt());
-        let q = MPoint::new_unchecked(1.0, -1.0, 0.0, 3.0f64.sqrt());
+        let p = MPoint::new_unchecked(-1.0, -1.0, 0.0, 3.0f32.sqrt());
+        let q = MPoint::new_unchecked(1.0, -1.0, 0.0, 3.0f32.sqrt());
         assert_abs_diff_eq!(p.distance(&q), q.distance(&p));
     }
 
     #[test]
     fn midpoint_distance() {
-        let p = MPoint::new_unchecked(-1.0, -1.0, 0.0, 3.0f64.sqrt());
-        let q = MPoint::new_unchecked(1.0, -1.0, 0.0, 3.0f64.sqrt());
+        let p = MPoint::new_unchecked(-1.0, -1.0, 0.0, 3.0f32.sqrt());
+        let q = MPoint::new_unchecked(1.0, -1.0, 0.0, 3.0f32.sqrt());
         let m = p.midpoint(&q);
         assert_abs_diff_eq!(p.distance(&m), m.distance(&q), epsilon = 1e-5);
         assert_abs_diff_eq!(p.distance(&m) * 2.0, p.distance(&q), epsilon = 1e-5);
